Moved subscription/activity roles in to a separate modul.
And make them translatable.
