+++ opengever/task/tests/test_activities.py
@@ -7,9 +7,9 @@ from ftw.testing.mailing import Mailing
 from opengever.activity import notification_center
 from opengever.activity.hooks import insert_notification_defaults
 from opengever.activity.model import Activity
-from opengever.activity.model.subscription import TASK_ISSUER_ROLE
-from opengever.activity.model.subscription import TASK_RESPONSIBLE_ROLE
-from opengever.activity.model.subscription import WATCHER_ROLE
+from opengever.activity.roles import TASK_ISSUER_ROLE
+from opengever.activity.roles import TASK_RESPONSIBLE_ROLE
+from opengever.activity.roles import WATCHER_ROLE
 from opengever.core.testing import OPENGEVER_FUNCTIONAL_ACTIVITY_LAYER
 from opengever.task.browser.accept.utils import accept_task_with_successor
 from opengever.testing import FunctionalTestCase
